requirements.txt
APScheduler==3.6.0
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.6.1
OpenCC==0.2
Pillow==6.0.0
pyaes==1.6.1
pyAesCrypt==0.4.3
pycparser==2.19
Pyrogram==0.13.0
PySocks==1.7.0
pytz==2019.1
six==1.12.0
TgCrypto==1.1.1
tzlocal==1.5.1
